SvelteKit upgrade
