+++ package.json
@@ -1,41 +1,36 @@
 {
-	"name": "spiceyourresume",
+	"name": "vite-project",
 	"version": "0.0.1",
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
-		"package": "vite package",
+		"package": "svelte-kit package",
 		"preview": "vite preview",
-		"prepare": "vite sync",
+		"prepare": "svelte-kit sync",
 		"check": "svelte-check --tsconfig ./jsconfig.json",
 		"check:watch": "svelte-check --tsconfig ./jsconfig.json --watch",
-		"lint": "prettier --ignore-path .gitignore --check --plugin-search-dir=. . && eslint --ignore-path .gitignore .",
-		"format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. ."
+		"lint": "prettier --check --plugin-search-dir=. . && eslint .",
+		"format": "prettier --write --plugin-search-dir=. ."
 	},
 	"devDependencies": {
+		"@popperjs/core": "^2.11.5",
 		"@sveltejs/adapter-auto": "next",
-		"@sveltejs/adapter-node": "^1.0.0-next.75",
 		"@sveltejs/kit": "next",
-		"@types/simple-oauth2": "^4.1.1",
-		"bootstrap": "^5.1.3",
-		"eslint": "^7.32.0",
+		"bootstrap": "^5.2.0",
+		"eslint": "^8.16.0",
 		"eslint-config-prettier": "^8.3.0",
-		"eslint-plugin-svelte3": "^3.2.1",
-		"prettier": "^2.5.1",
-		"prettier-plugin-svelte": "^2.5.0",
-		"sass": "^1.53.0",
+		"eslint-plugin-svelte3": "^4.0.0",
+		"prettier": "^2.6.2",
+		"prettier-plugin-svelte": "^2.7.0",
+		"sass": "^1.54.0",
 		"svelte": "^3.44.0",
-		"svelte-check": "^2.2.6",
+		"svelte-check": "^2.7.1",
 		"svelte-preprocess": "^4.10.7",
-		"typescript": "~4.6.2"
+		"typescript": "^4.7.4",
+		"vite": "^3.0.0"
 	},
 	"type": "module",
 	"dependencies": {
-		"@supabase/supabase-js": "^1.35.4",
-		"@sveltejs/adapter-static": "^1.0.0-next.38",
-		"@sveltejs/adapter-vercel": "^1.0.0-next.66",
-		"cookie": "^0.5.0",
-		"node-fetch": "^3.2.4",
-		"vite": "^3.0.3"
+		"@supabase/supabase-js": "^1.35.4"
 	}
 }
