+++ src/app.d.ts
@@ -2,6 +2,7 @@
 
 // See https://kit.svelte.dev/docs/types#app
 // for information about these interfaces
+// and what to do when importing types
 declare namespace App {
 	// interface Locals {}
 	// interface Platform {}
