+++ src/routes/__layout.svelte
@@ -1,5 +1,4 @@
 <script>
-	import '../app.scss';
 	import { onMount } from 'svelte';
 	import Navbar from '../components/Navbar.svelte';
 	import '../app.scss';
@@ -27,5 +26,5 @@
 	});
 </script>
 
-<navbar />
+<Navbar />
 <slot />
