+++ src/variables.scss
@@ -1,6 +1,6 @@
 /* Variables and mixins declared here will be available in all other SCSS files */
-/* Variables and mixins declared here will be available in all other SCSS files */
 @import 'bootstrap/scss/functions'; /* Override Bootstrap's variables here */
 @import 'bootstrap/scss/variables';
-@import 'bootstrap/scss/mixins';
 @import 'bootstrap/scss/maps';
+@import 'bootstrap/scss/mixins';
+@import 'bootstrap/scss/utilities';
