+++ svelte.config.js
@@ -1,9 +1,10 @@
 import preprocess from 'svelte-preprocess';
 import adapter from '@sveltejs/adapter-auto';
 
-export default {
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
 	kit: {
-		adapter: adapter({})
+		adapter: adapter()
 	},
 
 	preprocess: [
@@ -14,3 +15,5 @@ export default {
 		})
 	]
 };
+
+export default config;
