+++ vite.config.js
@@ -3,6 +3,7 @@ import { sveltekit } from '@sveltejs/kit/vite';
 /** @type {import('vite').UserConfig} */
 const config = {
 	plugins: [sveltekit()],
+
 	css: {
 		preprocessorOptions: {
 			scss: {
